+++ src/js/main.js
@@ -1,4 +1,4 @@
-const { ipcRenderer } = require("electron");
+const { ipcRenderer, shell } = require("electron");
 const fs = require("fs");
 const fsp = require("fs").promises;
 const path = require("path");